+++ further-python/shopping-cart/payment.py
@@ -1,9 +1,10 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
 class Payment(ABC):
     @abstractmethod
-    def pay(self, amount: float) -> str:
+    def pay(self, amount: Decimal) -> str:
         pass
